Hash rate graph (#15)
* WIP - simple bar graph]

* Bar lines WIP

* Graph wip

* WIP - Simple bar graph

* Review updates

* Trying to get data points to work

* Some clean up and refactoriing before Push

* PLaying with graph data
