+++ src/components/BlockExplorer.css
@@ -1,7 +1,13 @@
 .BlockExplorer {
-  align-items: center;
-  display: flex;
-  height: 100%;
-  justify-content: center;
-  flex-direction: column;
+    align-items: center;
+    display: flex;
+    height: 100%;
+    justify-content: center;
+    flex-direction: column;
+}
+.twoCol {
+    display: flex;
+    justify-content: space-around;
+    flex-direction: row;
+    width: 100%;
 }
